AG-14805 Label docs
