+++ packages/ag-charts-website/src/content/docs/axes-labels/_examples/axis-label-format/index.html
@@ -1,7 +1,1 @@
-<div class="example-controls">
-    <div class="controls-row">
-        <button onclick="zoomOut()">Zoom All Data</button>
-        <button onclick="zoomMonth()">Zoom Last Month</button>
-    </div>
-</div>
 <div id="myChart"></div>
